Update patch set 2
Patch Set 2: Verified+1

Thanks Ondrej for this significant contribution! As such, it will need a review by the Eclipse IP Team before we are able to accept it.

I have performed a code review and manually tested the implemented functionality. It looks good enough and works as expected. It is my understanding (from a mail exchange with Ondrej) that this is just an initial commit and further polish of the code will be done via separate contributions as circumstances allow.

I'm happy to give it a Verified +1 and file a CQ to initiate due diligence. Being the first ever significant contribution to the project, it represents an important milestone.

Thank you!

Patch-set: 2
Reviewer: Gerrit User 7131 <7131@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1